+++ src/utils/processStyleLoaders.js
@@ -2,6 +2,18 @@
 
 import escapeRegExp from 'escape-regexp';
 
+// Ensures '-loader' suffix for loaders in config for webpack compatibility
+const ensureLoadersSuffix = loadersArray => {
+  const loaderSuffixRegExp = new RegExp('^style-loader.*$|^css-loader.*$|^postcss-loader.*$|^sass-loader.*$|^resolve-url-loader.*$');
+  const suffixReplaceRegExp = new RegExp('^style|^css|^postcss|^sass|^resolve-url');
+  return loadersArray.map(loader => {
+    if (!loaderSuffixRegExp.test(loader)) {
+      return loader.replace(suffixReplaceRegExp, match => `${match}-loader`);
+    }
+    return loader;
+  });
+};
+
 /**
  * Injects 'resolve-url-loader' and 'sourceMap' param for 'sass-loader'
  *
@@ -16,8 +28,10 @@ Default is ['style', 'css', 'sass']
     `);
   }
 
+  const loadersWithSuffix = ensureLoadersSuffix(loaders);
+
   if (disableSassSourceMap && disableResolveUrlLoader) {
-    return loaders;
+    return loadersWithSuffix;
   }
 
   // We need to match user loaders in different formats:
@@ -29,9 +43,9 @@ Default is ['style', 'css', 'sass']
 
   const sassLoaderRegExp = getLoaderRegExp('sass');
   const sassLoader = (
-    loaders.find(loader => sassLoaderRegExp.test(loader))
+    loadersWithSuffix.find(loader => sassLoaderRegExp.test(loader))
   );
-  const sassLoaderIndex = loaders.indexOf(sassLoader);
+  const sassLoaderIndex = loadersWithSuffix.indexOf(sassLoader);
 
   if (!disableSassSourceMap) {
     if (!sassLoader) {
@@ -58,19 +72,19 @@ Default is ['style', 'css', 'sass']
 
 
     // eslint-disable-next-line no-param-reassign
-    loaders[sassLoaderIndex] = sassLoaderWithSourceMap;
+    loadersWithSuffix[sassLoaderIndex] = sassLoaderWithSourceMap;
   }
 
   if (!disableResolveUrlLoader) {
     const resolveUrlLoaderRegExp = getLoaderRegExp('resolve-url');
     const resolveUrlLoader = (
-      loaders.find(loader => resolveUrlLoaderRegExp.test(loader))
+      loadersWithSuffix.find(loader => resolveUrlLoaderRegExp.test(loader))
     );
 
     if (!resolveUrlLoader) {
-      loaders.splice(sassLoaderIndex, 0, 'resolve-url');
+      loadersWithSuffix.splice(sassLoaderIndex, 0, 'resolve-url-loader');
     }
   }
 
-  return loaders;
+  return loadersWithSuffix;
 }