+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. Items under
 Contributors: please follow the recommendations outlined at [keepachangelog.com](http://keepachangelog.com/). Please use the existing headings and styling as a guide, and add a link for the version diff at the bottom of the file. Also, please update the `Unreleased` link to compare to the latest release version.
 
 ## [Unreleased]
+##### Added
+- Ensures `-loader` suffix automatically for style loaders to safely keep compatibility with old Webpack versions as well as new versions with the suffix requirement from Webpack v2.1.0-beta.26 and onward. [#205](https://github.com/shakacode/bootstrap-loader/pull/205) by [kevinzwhuang](https://github.com/kevinzwhuang).
 
 ## [2.0.0.beta.14] - 2016-11-14
 ##### Added